requirements.txt
discord.py
python-dotenv
pytz
youtube_dl
neuralintents